refactor(storybook): update ThemeItem story and revise overall stories
