+++ src/components/features/Order/Payment/ReceiptForm/index.stories.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import { Meta, StoryObj } from '@storybook/react';
 import { ChakraProvider } from '@chakra-ui/react';
+import { useForm, FormProvider } from 'react-hook-form';
+import { OrderDataFormValues } from '@pages/Order';
 import GlobalStyles from '@assets/styles';
 import ReceiptForm from '.';
 
@@ -9,12 +11,25 @@ const meta: Meta<typeof ReceiptForm> = {
   component: ReceiptForm,
   tags: ['autodocs'],
   decorators: [
-    (Story) => (
-      <ChakraProvider>
-        <GlobalStyles />
-        <Story />
-      </ChakraProvider>
-    ),
+    (Story) => {
+      const methods = useForm<OrderDataFormValues>({
+        defaultValues: {
+          message: '',
+          hasCashReceipt: false,
+          cashReceiptType: '개인소득공제',
+          cashReceiptNumber: '',
+        },
+      });
+
+      return (
+        <ChakraProvider>
+          <FormProvider {...methods}>
+            <GlobalStyles />
+            <Story />
+          </FormProvider>
+        </ChakraProvider>
+      );
+    },
   ],
 };
